bsp: Support added in MB BSP for msrset and msrclr instructions
This patch fixes the CR#764881. It adds support for msrset and
msrclr instructions in several assembly files for cache, interrupt,
exception handling. It renames all *.s files in the MB BSP to *.S
so as to use include xparameters.h and use some of the MB settings
exposed through xparameters.h

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>