dptx: Broadcast messages need to have an LCR of 6.
An LCR (link count remaining) value of 6 is used to protect against loops.
If the LCR isn't processed at each node and attempted to be broadcasted on all
ports, waiting for a reply, a loop in the topology will result in the broadcast
message being forwarded infinitely.
Because of this, an LCR of 0 will prevent the message from being forwarded to
subsequent DP nodes.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>