sw_apps:zynq_fsbl: Fix for the library information in the FSBL template
In SDK GUI, the information of requirement of libraries for creating
Zynq FSBL was showing incorrect messages in Application project window.
When both xilffs and xilrsa libraries were absent, the error message
in SDK mentions only about xilffs. This is  corrected and the
error message now reflects the exact essential libraries that are missing.
This is fix for CR#837141.

Signed-off-by: Sarat Chand Savitala <saratcha@xilinx.com>