dp: Allow differentiate between XDptx and XDprx structures from a void *.
The XDp_Config structure is moved to the top of the instance structure for
XDptx. The XDprx structure, once implemented, will keep the XDp_Config structure
as the first member.

This allows functions to accept a void * argument and determine whether the
pointed structure is of type XDptx or XDprx by accessing the XDp_Config member's
IsRx member.

The reason this works is because XDp_Config is the first member in the
structure's order so the configuration structure will be placed in the same
relative location for both the XDptx and XDprx structures.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>