dp: XDptx->XDp and XDprx->XDp.
XDp is used as an argument for all API functions. If a user application uses a
deprecated structure, it is automatically converted to the top-level XDp
structure which contains either a XDp_Tx or XDp_Rx structure which corresponds
to the previous XDptx and XDprx.
This is done for back-wards compatability with the dptx driver to minimize user
effort in migrating to the new dp driver.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>