lwip: Fix issue on Rx path in hadling packets with Xilkernel
This patch fixes rx packet handling issue with Xilkernel. The issue
was processing a single packet from the input queue everytime the
semaphore is released in the ISR. Now, we iterate till the queue
is completely emptied. After the ISR releases the semaphore and before
the queue is read, there could be new packets in the queue and hence can
lead to situations where there could be considerable delay in processing
a particular packet.
This patch also ensures that the ISR releases the semaphore once for
an entry into Rx ISR instead of each received packet. This enhances
performance.

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>