dp: rx: Capture interrupt status from INTERRUPT_CAUSE bits 8 to 31.
Previously, INTERRUPT_CAUSE bits 8 to 31 were being lost due to the variables
assigned having type u8.

This fix now uses masking on the interrupts as the condition to run an interrupt
handler rather than assigning the mask result to a new variable and using the
variable as a boolean condition.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>