qspipsu: Switch to I/O mode before clearing RX FIFO
There is a bug wherein the DMA listening to RX empty status goes busy
if RX FIFO clear bit is set in the FIFO control register, even if there
is no transfer request. So switch to I/O mode always to clear RX FIFO and
restore the mode in the end.

Signed-off-by: Harini Katakam <harinik@xilinx.com>