qspi: Fix dummy mode in examples
QSPI HW doesn't actually care what the SPI mode is for the dummy phase
of flash commands. We have confirmed this with randomized testing.
However due to a core limitation of QEMU it is expected to match the
mode of the address phase for QEMU.

So since QEMU is the only platform that cares, set the dummy mode to
QEMUs expectation.

Signed-off-by: Peter Crosthwaite <peter.crosthwaite@xilinx.com>