dp: rx: mst: Added ability to set an I2C map for each port.
The "I2C map" terminology is used to refer to a set of I2C entries for a given
port. Each I2C entry contains an I2C address associated with that entry,
alongside a user-defined data structure which is to be pointed to.

This is the mechanism for which the driver allows the RX application to control
what data to be sent for a given I2C address when a REMOTE_I2C_READ is issued
by an upstream device.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>