dp: rx: mst: Keep the internal port number the same as the index.
When setting the configuration of a port, always keep the port number of the
port structure to the same value as the index of the structure's index in the
port structure array.

This simplifies functions that operate on the port structure since they don't
need to search each port structure in the array for the port number, the port
number is known from the index.

Note: The existence of the port number member in the port structure is required
since the TX populates this structure when receiving LINK_ADDRESS sideband
messages.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>