lib:xilpm:example: Use TCM-A for saving context
swbeta2 commit 90e16f97eb510ea91702729bf38bc7c7d5d62dba

FSBL inits the TCMs that are used by R5 Apps. We have vectors in TCM-A
and that  data is passed on to FSBL and it is initialized. So using the
same TCM block as vectors ensures that we use an intialized memory and
avoids ECC errors due to RMW or Reads of uninitialized memory locations.
In JTAG mode, TCM still needs to be initialized using XSDB.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>