PMUFW: Release RPU island reset on power up
swbeta2 commit 572db0eb0f48b4f7f5684abea721c6fac92ccdee

When powering up the RPU island the individual RPU core resets as well
as the reset for the whole island are asserted. To ensure proper resume,
the island reset needs to be released when the island is powered up.

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>