dp: rx: mst: Set number of ports in link address struct when exposing a port.
Previously, the number of ports was only being set when setting an input port.
If there are multiple input ports, the number of ports wasn't being accurately
reflected.

The user application was required to update the number of ports when setting a
downstream port.
This is now done within the driver.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>