nandpsu_v1_0: Fixed triggering of assert call when block reaches max block.
Since the check for the block to have reached the end block was at the end of the loop,
thus assert call triggering could happen if the block crosses the max block limit,
making the driver to hang. Thus to fix the triggering of assert call,
the check for the block to have reached the max block is being done at the start of
the loop in XNandPsu_CalculateLength function.
The u32 typecasting for the variable OffsetVar has been removed and the data type for
the variable TempLen has been changed to u64 as for higher block numbers the value for
these variables will require u64 datatype otherwise corrupted value will get stored
which will lead to wrong calculation of block.

Signed-off-by: Shakti Bhatnagar <shaktib@xilinx.com>