cpu_v2_3: Generate compiler flags for psu_microblaze
Compiler flags for psu_microblaze BSP, based on HW IP parameters, are not being generated by HSI. This is being done for normal microblaze by an unknown entity in the build flow and the same is being figured out. However the same flags can be generated by using cpu tcl and this work-around is implemented here.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>