llfifo: Update Register offsets in AXI4 data path as per latest IP version
This patch updates the register offsets in the AXI4 data path
as per latest IP version(v4.1).
The addresses are changed to accommodate increased data width.
With old address map and increased data width user had to generate AXI4 unaligned transactions.
Therefore, the address map was changed for ease of use in the IP.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>