nandpsu: Decrease the XNANDPSU_MAX_BLOCKS value
This change is to reduce the size of the static bbt table size
from 8KB to 4KB because so far we have not identified the
flash part that has more than 16K blocks and also it will
reduce the bsp size.

Driver warns if the device has more number of blocks than the
defined value so that this can be incremented in future and if
there is a part available.

Signed-off-by: Punnaiah Choudary Kalluri <punnaia@xilinx.com>
Acked-by: Anirudha Sarangi   <anirudh@xilinx.com>