PMUFW: IPI: Add IPI Framework
Basic APIs are provided for modules to send/receive IPI messages.
These are wrappers around IPI driver functions and are intended to
support a common message structure. Currently only a module's IPI ID
is considered.This can be extended to include checksum and other fields in
IPI message and dispatching IPIs based on module id.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>