hdcp1x: Converged CfgInitialize to one top-level.
Initialization functions for the cipher, port, and directions are now invoked
from within the XHdcp1x_CfgInitialize function directly.

No need to do check for inconsistencies between HW and SW for protocol and
direction because this scenario will never happen unless SDK generates the
xparameters.h incorrectly. The self test is the more appropriate place for this
check.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-by: Shadul Shaikh <shaduls@xilinx.com>