PMUFW: lscript: Force generation of single loadable section
Bootgen currently has no support for multiple loadable sections
in elfs. So a workaround is being implemented in PMUFW linker
script. All sections are forced to be loadable and dummy bytes
are added to fill the gaps. This change causes linker to throw
a warning about bss section being made PROGBITS.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>