PMUFW: pm: Implemented PmSetMaxLatency API, Latency accounting
Using PmSetMaxLatency one can assert that a slave node must be able
to return to its active state with a maximum latency. This latency
is also taken into account when changing a slave node's state.
Also, a call to PmSetMaxLatency may trigger a state change for the
slave node in question, if the current state does not meet the latency
requirement.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Signed-off-by: Davorin Mista <davorin.mista@aggios.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>