tmrctr: Updated initialization API.
Added configuration initialization and hardware initialization API
functions.
The XTmrCtr driver is now consistent with other drivers.

The previous initialization structure disallowed using a different
base address from what was found in the xparameters and the
configuration table of the core.
The XTmrCtr_Initialize function was overriding the effective base
address internal to the core - it still as to keep the API
functionality intact for backward-compatibility.
Internally, the function is calling the new APIs.

This poses an issue if the memory addresses have been translated
differently from the physical base address.
Also, such as is for subsystem drivers, the base address of one
of the subsystem's subcores will appear as a relative offset to
the base address of the subsystem.

This issue is addressed by providing the higher level with control
of which base address will be used by the driver.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-by: Shadul Shaikh <shaduls@xilinx.com>