hdcp1x: Updated configuration initialization.
The configuration initialization function is now consistent with
other drivers.

Added effective address as an argument to CfgInitialize.
This is necessary when memory addresses are translated differently
from the physical base address and for subsystem drivers whose
subcore base addresses appear as relative offsets to the base
address of the subsystem.

Moved configuration related content from LookupConfig to
CfgInitialize.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-by: Shadul Shaikh <shaduls@xilinx.com>