tmrctr: Protect against driver inclusion without core.
A driver may be included even when there is no associated core
present. (See the DEPENDS option in data/*.mdd).

If another driver includes this TmrCtr driver, despite no
associated timer counter core existing, then xtmrctr.tcl won't be
invoked by SDK and the xtmrctr_g.c will not be generated.
If this file is not generated, the BSP will include the default
xtmrctr_g.c which was previously assuming that there is at least
one instance of the core in the system.

Also, the XPAR_XTMRCTR_NUM_INSTANCES wasn't being defined in this
case.

Protection against such a scenario is required.
Some subsystem drivers will include all subcore drivers,
regardless whether they are a part of the system or not.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-by: Shadul Shaikh <shaduls@xilinx.com>