dp: tx: Added callback for MSA updates.
When the MSA is updated, provide a mechanism to run a user-defined
callback instead of using the MSA values from the driver's
structure.

This is useful if another DP core exists in the system (RX). Using
the new callback mechanism, the user can specify to use whatever
MSA values exist in the RX core as the values to transmit.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-by: Shadul Shaikh <shaduls@xilinx.com>