sw_services:xilsecure: Fix for authentication failures with larger partitions
It is observed that while authenticating larger bitstreams, FSBL fails.
Though there is wait for SHA3 done, CSU DMA done also needs to be
waited for and the completion of DMA transfer needs to be acknowledged
by clearing DONE status.
This will make sure that the next transfer doesn’t assume DONE when it isn’t.

Signed-off-by: Sarat Chand Savitala <saratcha@xilinx.com>
Acked-by: Bhavik Ameta <bameta@xilinx.com>