xilpm: Added XPm_RegisterNotifier PM API call
The register notifier API call in xilpm now requires passing a pointer
to a XPm_Notifier struct, which contains the node ID and event ID the
notifications are requested for.

The new files pm_callback.h/c contain the functions for managing
the list of notifiers, supporting multiple pending notifications
at once, but also user-defined callback functions to avoid polling.

Signed-off-by: Filip Drazic <filip.drazic@aggios.com>
Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Signed-off-by: Davorin Mista <davorin.mista@aggios.com>
Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Acked-by: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>