PMUFW: resets: Work-around for AIB ACK signal glitch
AIB ACK signal to PMU may have a glitch due to a bug in 1.0 Silicon.
So we confirm the ACK again to reduce the probability of false ACKs.
This additonal check may be retained for 2.0 and above. It provides
an extra level of protection while not impacting the performance, when
compared to adding a silicon version check.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>