fifo: Fix Incorrect AXI4 Base address being exported
AXI FIFO Can be configured in two interfaces for data path
AXI4/AXI4-Lite. When fifo is configured axi4-lite mode
The AXI4 base address got exported is to the hdf is a defalut
value from the IP(0x80001000).But the AXI4 baseaddress is nowhere
visiable in the design which is confusing the user.

This patch fixes this issue by exporting zero to the
AXI4 base address when fifo is configured in AXI4-Lite mode.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Anirudha Sarangi <anirudh@xilinx.com>