dp: tx: Disable line reset for reduced blanking resolutions.
When enabled, the end of line reset generated by the TX framing
logic to the video interface FIFO clears any stalled video data
stored in the FIFO.
This is enabled by default in the IP.

The end of line reset happens at the end of the horizontal
active period and start of blanking.
Reduced blanking resolutions have a shorter blanking interval,
hence there exists a narrower window which the reset can happen.

Due to there existing a latency from the framing logic to the
video FIFO, the line reset was clearing the next active line
of video data for resolutions with reduced blanking.

To handle this, when the DP TX driver detects that a reduced
blanking resolutions is being sent, the end of line reset is
disabled. The stalled data is cleared by the end of frame
reset.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-by: Kapil Usgaonkar <kapilu@xilinx.com>