PMUFW: pit: Fix PIT clock frequency value
PITs run on the PMU clock which in turn could be PS_REF_CLK or IRO.
For 1.0, this is mostly limited to PS_REF_CLK due to issues with IRO.
However, for production silicon, PMU would be clocked ONLY from IRO.
Taking into account the above scenarios, clock frequency setting is
moved into config.h file and can be changed based on the user setup.
In future, this may be fixed to IRO frequency, probably by checking
PS version.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-by: Soren Brinkmann <soren.brinkmann@xilinx.com>