sw_apps:zynqmp_fsbl: Setting memory attributes for DDR - A53
Routines implemented to set memory attributes for whole
DDR memory. If FSBL uses the one in bsp which flushes
data cache every time it is called, it consumes more time.
Hence, routines are implemented in FSBL to set the attaributes
for whole DDR and data cache will be flushed only once at the end.

Signed-off-by: Ramakrishna Ganeshu Poolla <rpoolla@xilinx.com>
Acked-by: Sarat Chand Savitala <saratcha@xilinx.com>