bsp: a9: remove setting for SMP bit in boot code
This patch removes the setting of SMP bit for cortex-a9 in boot code
as standalone application does not run as SMP and modified memory
attribute for DDR to non-shareable since processor is in non-SMP mode.
This was a legacy code and should be removed as per arm specifications.
It may result in improved cache behavior

Signed-off-by: Kinjal Pravinbhai Patel <patelki@xilinx.com>
Acked-by: Anirudha Sarangi   <anirudh@xilinx.com>