drivers: ttc: modified ttc testapp tcl to avoid compilation error
This patch modifies ttc testapp tcl to generate proper ttc
IDs for peripheral test when some of the ttcs(not all) are
present in design. E.g. if only ttc1 is present the ID should
be *_TTC_0_* instead of *_TTC_3_*. This patch also modifies
the tcl to exclude ttc3 in perihperal test for psu_cortexr5
for zynqmp soc as TTC3 is being used by sleep APIs in the BSP

Signed-off-by: Kinjal Pravinbhai Patel <patelki@xilinx.com>
Acked-by: Anirudha Sarangi <anirudh@xilinx.com>