cpu_cortexa9: Update to fix issues related to toolchain name
This patch updates the cortexa9 cpu tcl to consider only the toolchain
name when a complete path for the toolchain is passed. This patch also
makes updates to have explicit separate cases for code sourcery and
armcc toolchains instead of clubbing these two in a single "else" case.

The changes ensure that when toolchain passed has the complete path, the
conditional checks only consider the toolchain name from the path.
Without this change when a complete path is passed, the tcl misbehaved.
Having explicit cases for armcc and code sourcery ensure that if ever a
new toolchain is passed(not supported by xilinx) the
extra_compiler_flags is not updated with wrong parameters.

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>
Acked-by: Kinjal Pravinbhai Patel <patelki@xilinx.com>