sw_apps:zynqmp_fsbl: Workaround in R5 FSBL while validating partition
FSBL only supports LOVEC and user app need to use BSP built for LOVEC.
This means when FSBL runs on R5, FSBL's vectors are overwritten with
those of uesr applications. To avoid happening of this, while copying
partitions, application partition's memory region overlapping with
this vectors are saved in a local buffer and are actually copied to
memory during handoff (i.e. copying on vectors region is skipped
during actual partition copy).

However, in above case, during authentcation and decryption of
partition, mismatch occurs for the bytes which are skipped copying.

Hence, in this patch,  as a workaround the above mentioned skip is
not beinge done when partiion has to be either authenticated or
decrypted.

Known issue with this change is that, when FSBL runs on R5 and user
partition has to be either authenticated or decrypted, in such cases
occurence of any exception, it will not be reported.

A proper and complete fix, to happen later, will fix this.

Signed-off-by: Sarat Chand Savitala <saratcha@xilinx.com>
Acked-by: Krishna Chaitanya Patakamuri <kpataka@xilinx.com>