sw_apps: zynqmp_fsbl: Fix Perf Measurement related build error
The XTime type is not defined in R5 builds in case
SLEEP_TIMER_BASEADDR is not defined. This results in compilation to
fail. Fix this by including the function declaration using XTime type
under the XFSBL_DEBUG condition.

Also change the earlier macro FSBL_PERF to XFSBL_PERF -- this is in
line with the other conditional macros defined in the fsbl code.

Define XFSBL_PERF for R5 only in case of SLEEP_TIMER_BASEADDR being
defined.

Signed-off-by: Sughosh Ganu <sughoshn@xilinx.com>
Acked-by: Sarat Chand Savitala <saratcha@xilinx.com>