bsp: a9: Modified translation table to fix ACP coherency issue
This patch reverts the changes to run the standalone code in
non-SMP mode. The ACP port needs the memory to be marked as
shareable to snop through caches. In non-SMP mode if the memory
is marked as shareable, it bypasses the L1 cache and put the data
directly to L2 cache which may impact the performance. To fix the
ACP coherency problem, this patch modifies the boot code to keep
the processor in SMP mode and translation_table.s to mark the memory
to be shareable

Signed-off-by: Kinjal Pravinbhai Patel <patelki@xilinx.com>
Acked-by: Anirudha Sarangi   <anirudh@xilinx.com>