Iicps: Clear the hold bit in ZYNQ to get completion interrupt propertly.
This patch clear the hold bit when the receive count is less than fifo
depth and does so after initiating repeated start by writing slave
address.

Idealy this check should include !IsRepeatedStart but since
Zynq has the limitation of not allowing any transfers to
follow a read transfer with repeated start, this is unecessary.

Signed-off-by: Nava kishore Manne <navam@xilinx.com>
Acked-by: Harini Katakam <harinik@xilinx.com>