bsp: r5: modified boot code to enable comparators for non-JTAG boot mode
This patch modifies R5 boot code to enable fault logs for only
non-JTAG boot modes in lock-step. The debug logic had been put to
reset to avoid interference for lock-step comparators. But that
required to take the debug logic out of reset to enable access for
debugger to perform any actions once the code is executed which
may result into unpredictable behavior. This patch also removes
the enabling back of debug logic in _exit.c.

Signed-off-by: Kinjal Pravinbhai Patel <patelki@xilinx.com>
Acked-by: Anirudha Sarangi <anirudh@xilinx.com>