csudma: Fix race condition in recv path
Existing driver doing a Invalidation of the recv
buffer during transfer.

If source and destination pointing to the same buffer
for this case the current driver will work for a53 but won't
work for r5. a53 Cache Invalidation API does both flush and invalidation.
But it is not the case for r5.

This patch fixes this issue for r5 case.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Anirudha Sarangi <anirudh@xilinx.com>