PMUFW: pm: Clear GIC Proxy interrupt when PmSetWakeupSource is called
Interrupt shall be cleared before it is enabled in GIC Proxy. If the
interrupt is cleared just before it is enabled (when FPD is powered off),
there is a time window between setting a PWRCTRL bit on APU side (after
that APU core is not receiving any interrupts) and the moment at which an
interrupt is cleared on PMU side. To avoid losing an interrupt which
occurrs in this time window, we clear interrupt at GIC Proxy when set
wake-up source PM API is called. The set wake-up source shall be called
before self suspend which sets the associated PWRCTRL bit for the APU core.
Accordingly, any occurrance of an interrupt in the time window should be
captured by the PMU.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>