PMUFW: pm: Clear GIC proxy group IRQ before enable
Clear potentially pending IRQs and ensure new edges are generated
towards the GPI module.

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>

[ MS
Group interrupt is now cleared and enabled only when the first interrupt
from the group is going to be enabled. Clearing a group interrupt every
time when an interrupt within the group is enabled may cause the loss of
interrupts.
]
Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>