scugic: Fix for Distributor Init issue for a particular R5 OPEN AMP Config
In a scenario where R5 is operating in split mode with one CPU in open amp config
with A53 and other running some baremetal app there is a situation  when the AMP
solution will fail to work. The current configuration assumes that the R5 running
baremetal app will initialize the Distributor and the one in AMP configuration
will just use it. But if for some reason, the baremetal app crashes before
initializing the Distributor, the whole AMP solution can fail.

The patch makes changes so that under OPEN_AMP flag (open amp use case for the
above R5 configuration), a check is first made to see if the Distributor is in
initialized state of not. If not, the AMP CPU goes ahead and initializes the
distributor assuming that the other CPU did not do the initialization.
Please note that, it is assumed that in a typical use case the baremetal app (non amp)
always comes up first and initializes the distributor.

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>
Acked by: Kinjal Pravinbhai Patel <patelki@xilinx.com>