bsp: r5: Skip clearing SBSS/BSS sections on warm boot
When processor boots the startup code unconditionaly clears BSS and SBSS
sections. For the warm boot, the sections shall not be cleared.
This patch introduces a check whether the boot is warm or cold so the
sections are cleared conditionally. Boot type is determined by reading
PWRCTRL register of the resuming CPU.

Signed-off-by: Filip Drazic <filip.drazic@aggios.com>
Acked-by: Anirudha Sarangi <anirudh@xilinx.com>