iic: Added low-level bus busy check.
XIic_CheckIsBusBusy is a low-level API which operates in the same
manner as XIic_IsBusBusy, except its argument is the base address
of the instance as opposed to a pointer to the instance.

XIic_IsBusBusy accepts an instance pointer as an argument and now
is a simple wrapper around XIic_CheckIsBusBusy.

Exposing a low-level API for this allows for other low-level APIs
to reuse the common bus busy check.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-for-series: Anirudha Sarangi <anirudh@xilinx.com>