sw_apps:zynqmp_fsbl: Added support for TCM ECC Initialization
This patch adds support for TCM ECC Initialization in FSBL. For R5, TCM ECC is
always initialized. For A53, by default TCM ECC is not initialized as this
also invloves powering up of RPU. However, provision is made in code for TCM
ECC Initialization even for the case of A53. This is achieved by making value
of FSBL_A53_TCM_ECC_EXCLUDE_VAL 0 in xfsbl_config.h (which is 1 by default).
Also, clean up of code done involving usage of TCM/ATCM/BTCM.

Signed-off-by: Sarat Chand Savitala <saratcha@xilinx.com>
Acked-by: Krishna Chaitanya Patakamuri <kpataka@xilinx.com>