PMUFW: pm: core: Set wake-up source must be applicable to all PUs
The set wake-up source is implemented as it would be called only
from APU (calling this API makes sense only for APU). However,
for other PUs it is considered valid to call this API as well, so
the implementation must gracefully handle all scenarios. If the
target is not APU, just return success (RPU_0/1 are always
woken-up through GIC wake and there is nothing that PMU can
configure/enable regarding the wake. The GIC wake interrupt is
enabled once the PU suspends)

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>