PMUFW: pm: Change the order of data definitions in pmReqData array
Due to the performance/memory reasons the elements from pmReqData
array are linked at the beginning of the master/slave's lists.
The resulting order of data definitions is thereby reversed compared
to the previous implementation. This patch reorders the definitions
of elements in pmReqData array to preserve the backward compatibility
(expected order of state transitions).

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>