PMUFW: pm: Process PmSetWakeUpSource for all masters
The request needs to be processed for all masters in order to properly
set the flags used for tracking the set wake-up source requests.
Compared to the previous implementation, if the slave has been set as
the wake source the PMU needs to make sure that the slave enters state
in which it has PM_CAP_WAKEUP capability (the PM_CAP_WAKEUP capability
doesn't have to be provided with the PmSetRequirement call, a master
doesn't have to take care about these dependencies).
If the call is not processed for all masters the PMU will not know that
there is such a dependency (master wake-up depends on slave's state).

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>