PMUFW: pm: In PmSetRequirement set the wake-up capability flag if needed
If the slave is set as the wake source prior to the PmSetRequirement
call, the PMU must ensure that the PM_CAP_WAKEUP flag is set if the
master does not specify that flag as the argument of the
PmSetRequirement. When this flag is set, the PMU will configure the
state in which the slave can trigger a wake-up interrupt.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>