sw_apps: zynqmp_fsbl: Add XFsbl_PmInit call
This function is used to notify PMU firmware that initialization of all
PM related registers is completed. This function is invoked from the
handoff stage.
Notification is performed by sending an IPI to the PMU if PMU firmware
is ready. However, in JTAG boot mode, PMU firmware can be loaded after
FSBL finishes execution, in which case IPI will not be sent.
To cover this situation, FSBL uses PMU_GLOBAL_PERS_GLOB_GEN_STORAGE5
register to mark that it has completed. PM firmware can read this register
to get information about FSBL status.

Signed-off-by: Filip Drazic <filip.drazic@aggios.com>
Acked-by: Sarat Chand Savitala <saratcha@xilinx.com>