qspipsu: Add PollData and PollTimeout support
The Generic Quad-SPI controllerr has a feature to read the status of SPI
device on its own. When a write/erase command issued to a flash device,
the software polls in multiple iterations to know the status of the device.
This can be done by the hardware using PollData and PollTimeout.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Acked-for-series: Punnaiah Choudary Kalluri<punnaia@xilinx.com>