PMUFW: AIB: Add AIB APIs
AXI Isolation Blocks are used to gracefully terminate and block
transactions on the AXI interconnect during power/reset events.
AIBs are placed at critical paths in the interconnect and can be
individually enabled or disabled as required. APIs to enable,
disable and check status of an AIB are provided.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-for-series: Sarat Chand Savitala <saratcha@xilinx.com>