dptxss: Support native video mode compilation.
In case of native video mode, there are no VTC instances
a part of the hierarchical IP (subsystem). In absence of
VTC instantiation, the vtc driver was not being included
into the BSP and compilation was failing.

This patch fixes compilation for such a scenario and VTC
functionality is excluded at runtime.

Signed-off-by: Andrei-Liviu Simion <andrei.simion@xilinx.com>
Acked-for-series: Anand Ashok Dumbre <anandash@xilinx.com>