PMUFW: pm: Added PmInit API call
The PmInit shall be called after the FSBL initializes hardware.
When the function is called, the PFW power management will read
the current configuration of the registers and update its data
structures accordingly.

Note: only the clock related code depends on such a initialization.
If the PmInit call is not made, the clock pointers will not be
initialized and the power management will not handle any clock
dependencies.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Sarat Chand Savitala <saratcha@xilinx.com>