qspipsu: Use lqspi in single mode
This patch changes the lqspi application to operate in singe mode.
In Dual parallel mode ZynqMP GQSPI writes data in bytes(Even bytes in lower flash and
odd bytes in upper flash), where as LQSPI reads the data in bitwise. So this is causing
data mismatch while reading. So proceed with single mode irrespective of flash connection.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Acked-by: Sai Krishna Potthuri <lakshmis@xilinx.com>