sw_apps: zynqmp_fsbl: Avoid multiple IPI triggers during PMUFW load
As per PMUFW load process, once IPI-0 is triggered and PMU goes into
sleep, FSBL can load all the partitions and wake up PMU. In FSBL, when
multiple partitions for PMUFW are present, IPI is being triggered for
every partition, which is not correct. So, we change the flow such that
IPI-0 is triggered only for the first PMUFW partition. A non-zero
execution address is considered as an indication for detecting first
PMUFW partition. We still check if PMU MB is in sleep or not before
copying each partition as an extra safegaurd.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-by: Krishna Chaitanya Patakamuri <kpataka@xilinx.com>